requirements.txt
alembic==1.8.1
aniso8601==9.0.1
apispec==5.2.2
click==8.1.3
Flask==2.2.2
flask-apispec==0.11.4
Flask-Migrate==3.1.0
Flask-RESTful==0.3.9
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.1
MarkupSafe==2.1.1
marshmallow==3.17.0
packaging==21.3
psycopg2-binary==2.9.3
PyJWT==2.4.0
pyparsing==3.0.9
pytz==2022.1
six==1.16.0
SQLAlchemy==1.4.40
webargs==8.2.0
Werkzeug==2.2.2
zipp==3.8.1
